+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Initial version of ESRI/ArcGIS geopicker (geopoint only). See [README.md](https://github.com/enketo/enketo-core/blob/master/src/widget/geo-esri/README.md.
+
 [4.11.0] - 2016-10-13
 ---------------------
 ##### Changed